refactor(user): separate the custom user manager from the model
separate the logic of the custom user manager from the model file
