+++ traceabilitymatrix/user/models.py
@@ -1,38 +1,9 @@
 from django.db import models
-from django.contrib.auth.models import AbstractUser, BaseUserManager
+from django.contrib.auth.models import AbstractUser
+from .managers import UserManager
+from django.contrib.auth.models import PermissionsMixin
 
-
-class UserManager(BaseUserManager):
-
-    def _create_user(self, email, password=None, **extra_fields):
-        """Create and save a User with the given email and password."""
-        if not email:
-            raise ValueError("The given email must be set")
-        email = self.normalize_email(email)
-        user = self.model(email=email, **extra_fields)
-        user.set_password(password)
-        user.save(using=self._db)
-        return user
-
-    def create_user(self, email, password=None, **extra_fields):
-        extra_fields.setdefault("is_staff", False)
-        extra_fields.setdefault("is_superuser", False)
-        user = self._create_user(email, password, **extra_fields)
-        return user
-
-    def create_superuser(self, email, password=None, **extra_fields):
-        extra_fields.setdefault("is_staff", True)
-        extra_fields.setdefault("is_superuser", True)
-
-        if extra_fields.get("is_staff") is not True:
-            raise ValueError("Superuser must have is_staff=True.")
-        if extra_fields.get("is_superuser") is not True:
-            raise ValueError("Superuser must have is_superuser=True.")
-
-        return self._create_user(email, password, **extra_fields)
-
-
-class User(AbstractUser):
+class User(AbstractUser, PermissionsMixin):
     class Role(models.TextChoices):
         Admin = "Admin", "Admin"
         TeamMember = "Team Member", "Team Member"
